refactor: split daemon_ha_userspace.go by concern (#4659)
pkg/daemon/daemon_ha_userspace.go had grown to 1123 LOC interleaving
four distinct concerns of the userspace-dataplane HA control path: wire
delta<->session conversion, the delta stream/drain/queue loop, bulk
session export, and takeover/demotion readiness plus failover barriers.
codex-review-174 #17 flagged the mix as a real review hazard — the
ACK/withhold and demotion-barrier logic sits next to endian/NAT
conversion helpers, so a reviewer scanning the barrier path has to wade
through unrelated byte-order code.

This is a class-B HA control path, so the split is PURE code motion:
every top-level declaration is moved verbatim, byte-for-byte, with no
behavior change. Concerns are grouped into four sibling files in package
daemon:

  - daemon_ha_userspace_convert.go (16 decls): buildZoneIDs, the
    endian/NAT-port/MAC helpers, userspaceSessionFromDeltaV4/V6, and the
    forward-wire alias/key builders.
  - daemon_ha_userspace_stream.go (11 decls): the drainer and
    event-stream provider interfaces, shouldSyncUserspaceDelta,
    syncUserspaceSessionDeltas, runUserspaceEventStream,
    wireUserspaceEventStreamCallbacks, handleEventStreamDelta,
    handleEventStreamFullResync, eventStreamFallbackLoop,
    queueUserspaceSessionDeltas, drainUserspaceSessionDeltasWithConfig.
  - daemon_ha_userspace_export.go (4 decls): the exporter interfaces,
    primaryOwnerRGIDs, exportUserspaceOwnerRGSessionsWithConfig.
  - daemon_ha_userspace_readiness.go (18 decls): the mode / takeover /
    transfer-readiness / HA-protocol-mismatch interfaces, the
    demotion-prep suppression-window methods, the
    prepareUserspaceRGDemotionWithTimeout peer barrier, the
    manual-failover wrappers, and userspaceDataplaneActive /
    userspaceRGConfigured / checkUserspaceTakeoverReadiness.

daemon_ha_userspace.go stays as the shared shell (3 decls:
userspaceXSKBindingController, buildZoneRGMap, rgHasRETH).

Load-bearing semantics preserved byte-identical: the delta-slice
behavior, the ACK/withhold return contract of handleEventStreamDelta
(returns false only for transient readiness gaps so the helper replays),
the userspaceDeltaSyncMu (delta drain) vs userspaceDemotionPrepMu
(demotion suppression window) lock separation, and the failover
peer-barrier logic. The split moves methods, not the lock discipline.
Same package, so every unexported symbol stays reachable —
buildZoneIDs (consumed by stream + export), the exporter interfaces
(used by daemon_ha_sync.go), and buildZoneRGMap / rgHasRETH (used by
daemon_apply.go, daemon_ha_vip.go, and package tests).

Validation: verified byte-for-byte against origin/master with a
go/parser decl-diff over the union of the new files — 52 top-level
declarations, identical=52, 0 bodydiff / 0 missing / 0 extra (import
lists differ per file as expected and are excluded from the diff).
go build ./... clean, go vet ./pkg/daemon/... clean, and
go test ./pkg/daemon/... green (the userspace-sync and HA-sync semantic
tests are the gate — all pass). No design/doc change: pkg/daemon/
README.md carries no per-file layout entry for this file and the split
is behavior-identical.

Closes #4659
